krgsched: Introducing new notifiers for fork() and exit()
This patch adds new notifiers hook inside the fork() and
exit() syscall, in order to be able to catch process arrival and
death.

This is in the scope of a better migration tracking: not only
discover process when they migrate.

New notifiers (kmh_process_{birth,death}) are only called if the
process we catch is running in the Kerrighed container, in such a
way that only process from kerrighed container that are migrable
or distant fork()able will be considered.

By default, migrations issues a do_exit() with the notify parameter
set to false. We need to check for this parameter to ensure that
our Kerrighed process isn't in a migrating status. If we don't, we will
catch fake migrations: process seems to die and then reappear on
another node, leading to inconsistency in the graph.

[ Cleaned up a few coding styles issues -- Louis Rilling ]

Signed-off-by: Alexandre Lissy <alexandre.lissy@etu.univ-tours.fr>
Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>