krgsched: Introducing the ProcessSize probe.
This patch adds a new probe called processsize. It will be used by the
mattload probe which will be introduced later.

This probe currently has only one probe source, called total_vm, which
returns the value of mm->total_vm of a struct task_struct. The probe
source can be either fed with a pid_t parameter or it will return
(print) a list of process ids and total_vm values.

It allows to have an idea of the memory consummed by a process.

This is part of the API written by Mathieu Dabert.

Signed-off-by: Alexandre Lissy <alexandre.lissy@etu.univ-tours.fr>
Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>