krgclusterip: Introduce support for cluster IP
This commit introduces a Kerrighed way of defining a single IP address for
a cluster. The principle is to divert iptables' CLUSTERIP target when used
inside the Kerrighed container.

The IP socket address spaces of the Kerrighed containers of all nodes are
merged in the following way:
- a socket bound to a port on INADDR_ANY or on a cluster IP (that is setup
  with CLUSTERIP in the Kerrighed container) will reserve the port on all
  nodes. No other socket can be bound to the same port, whatever the node
  on which it is created.
- as soon as a cluster IP is setup on all nodes, above mentioned sockets
  will be able to receive network traffic.
- a socket bound to a port on a local IP (that is not setup with CLUSTERIP
  on this node) will keep the port available on other local IPs and other
  nodes.

Currently only UDP and TCP over IPv4 are supported.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>