[ENHANCEMENT] Add a KDDM object reference counter.
The KDDM object mapcount counts the number of data structures sharing
the object.

This new refcount counts the number of tasks having a reference to the
object. mapcount counts for 1 in the refcount, whatever its value if >
0.

This new counter is usefull to fix some corner cases in the object
removal case for instance.