[FEATURE] Add a lock free way to get a KDDM object.
This will be used by the removal KDDM code. Some actions need to be
carried about since KDDM are locked in the reconfiguration phaze.
Since KDDM are lock and cannot be created, removed or modified, we can
do operations without acquiring KDDM locks.

However, we still need to lock the object since several actions could
be done on the same object at the same time.