kddm_flush_set(): Allow to delay IO linker's remove_object() to sleepable context
It is not allowed to sleep inside __kddm_flush_object() (and thus in the
user-provided callback of kddm_flush_set()), so it is not safe to call
some IO linker's remove_object().

This commit introduces an additional argument to KDDM objects iterators so
that, while walking the local table, kddm_flush_set() can actually destroy
object entries  and collect missing calls to IO linker's remove_object().

The IO linker is then called once the object table walk is finished.

This logic is controlled by a new flag KDDM_NEED_SAFE_WALK that must be
specified when creating a KDDM set.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>