[BUGFIX] kddm_flush_set(): Fix broken consistency and prob owner chains
Commit f2f34433dea8ea3290ea769430287853411ac743
	"[BUGFIX] remove/kddm: Fix lock imbalance in [_]flush_kddm_set()"
opened another bug consisting in actually freeing parts of the object
tables, loosing references to object entries as well as breaking prob
owner chains, and also enabling the existence of several object owners.

This commit allows ->for_each_obj_entry() to distinguish between freed
object entries (that is the table can forget about it) and only cleared
object entries (that is the object entry was invalidated and unlocked, the
object was locally removed, but the object entry should remain).

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>